Move apps into their own packages
