+++ pkg/bindata/templates/downloaders/deluge.html
@@ -4,7 +4,7 @@
                                                     <div style="float: left;"><img src="{{files}}/images/logo/deluge.png" style="height:50px;">
                                                     </div><h2 class="mb-3" style="margin-bottom:-45px">Deluge</h2>
                                                     <div style="float: right;">
-                                                        <button id="download-Deluge-addbutton" onclick="addInstance('download', 'Deluge')" data-prefix="Apps" data-sslname="VerifySSL" data-names='["Name","","Config.URL","Config.Password","Interval","Timeout"]' type="button" title="Add another instance of Deluge." class="add-new-item-button btn btn-primary"><i class="fa fa-plus"></i></button>
+                                                        <button id="download-Deluge-addbutton" onclick="addInstance('download', 'Deluge')" data-prefix="Apps" data-sslname="ValidSSL" data-names='["Name","","Config.URL","Config.Password","Interval","Timeout"]' type="button" title="Add another instance of Deluge." class="add-new-item-button btn btn-primary"><i class="fa fa-plus"></i></button>
                                                     </div>
                                                 </td>
                                                 <td colspan="7" class="tablet-hide desktop-hide">
@@ -93,7 +93,7 @@
                                                                 {{- end}}
                                                                 <input type="text" id="Apps.Deluge.{{$index}}.Config.URL" onChange="showhttps($(this).val(), '#Deluge{{$index}}SSL');" name="Apps.Deluge.{{$index}}.Config.URL" data-index="{{$index}}" data-app="Deluge" class="client-parameter form-control input-sm" data-group="download" data-label="Deluge {{instance $index}} URL" data-original="{{$app.URL}}" value="{{$app.URL}}">
                                                                 <div style="width:30px; max-width:30px;{{if not (contains $app.URL "https://")}}display:none;{{end}}" id="Deluge{{$index}}SSL" class="input-group-addon input-sm">
-                                                                    <input type="checkbox" id="Apps.Deluge.{{$index}}.VerifySSL" name="Apps.Deluge.{{$index}}.VerifySSL" data-index="{{$index}}" data-app="Deluge" class="client-parameter" data-group="download" data-label="Deluge {{instance $index}} SSL" data-original="{{$app.VerifySSL}}" {{if $app.VerifySSL}}checked {{end}}value="true">
+                                                                    <input type="checkbox" id="Apps.Deluge.{{$index}}.ValidSSL" name="Apps.Deluge.{{$index}}.ValidSSL" data-index="{{$index}}" data-app="Deluge" class="client-parameter" data-group="download" data-label="Deluge {{instance $index}} SSL" data-original="{{$app.ValidSSL}}" {{if $app.ValidSSL}}checked {{end}}value="true">
                                                                 </div>
                                                             </div>
                                                         </div>
