+++ pkg/bindata/templates/downloaders/qbit.html
@@ -4,7 +4,7 @@
                                                     <div style="float: left;"><img src="{{files}}/images/logo/qbittorrent.png" style="height:50px;"></div>
                                                     <h2 class="mb-3" style="margin-bottom:-45px">qBittorrent</h2>
                                                     <div style="float: right;">
-                                                        <button id="download-Qbit-addbutton" onclick="addInstance('download', 'Qbit')" data-prefix="Apps" data-sslname="VerifySSL" data-names='["Name","URL","User","Pass","Interval","Timeout"]' type="button" title="Add another instance of Qbittorrent." class="add-new-item-button btn btn-primary"><i class="fa fa-plus"></i></button>
+                                                        <button id="download-Qbit-addbutton" onclick="addInstance('download', 'Qbit')" data-prefix="Apps" data-sslname="ValidSSL" data-names='["Name","URL","User","Pass","Interval","Timeout"]' type="button" title="Add another instance of Qbittorrent." class="add-new-item-button btn btn-primary"><i class="fa fa-plus"></i></button>
                                                     </div>
                                                 </td>
                                                 <td colspan="7" class="tablet-hide desktop-hide">
@@ -94,7 +94,7 @@
                                                                 {{- end}}
                                                                 <input type="text" id="Apps.Qbit.{{$index}}.URL" onChange="showhttps($(this).val(), '#Qbit{{$index}}SSL');" name="Apps.Qbit.{{$index}}.URL" data-index="{{$index}}" data-app="Qbit" class="client-parameter form-control input-sm" data-group="download" data-label="Qbit {{instance $index}} URL" data-original="{{$app.URL}}" value="{{$app.URL}}">
                                                                 <div style="width:30px; max-width:30px;{{if not (contains $app.URL "https://")}}display:none;{{end}}" id="Qbit{{$index}}SSL" class="input-group-addon input-sm">
-                                                                    <input type="checkbox" id="Apps.Qbit.{{$index}}.VerifySSL" name="Apps.Qbit.{{$index}}.VerifySSL" data-index="{{$index}}" data-app="Qbit" class="client-parameter" data-group="download" data-label="Qbit {{instance $index}} SSL" data-original="{{$app.VerifySSL}}" {{if $app.VerifySSL}}checked {{end}}value="true">
+                                                                    <input type="checkbox" id="Apps.Qbit.{{$index}}.ValidSSL" name="Apps.Qbit.{{$index}}.ValidSSL" data-index="{{$index}}" data-app="Qbit" class="client-parameter" data-group="download" data-label="Qbit {{instance $index}} SSL" data-original="{{$app.ValidSSL}}" {{if $app.ValidSSL}}checked {{end}}value="true">
                                                                 </div>
                                                             </div>
                                                         </div>
