+++ pkg/bindata/templates/downloaders/sabnzbd.html
@@ -4,7 +4,7 @@
                                                     <div style="float: left;"><img src="{{files}}/images/logo/sabnzbd.png" style="height:50px;"></div>
                                                     <h2 class="mb-3" style="margin-bottom:-45px">SabNZB</h2>
                                                     <div style="float: right;">
-                                                        <button id="download-SabNZB-addbutton" onclick="addInstance('download','SabNZB')" data-prefix="Apps" data-sslname="VerifySSL" data-names='["Name","", "URL","APIKey","Interval","Timeout"]' type="button" title="Add another instance of SabNZB." class="add-new-item-button btn btn-primary"><i class="fa fa-plus"></i></button>
+                                                        <button id="download-SabNZB-addbutton" onclick="addInstance('download','SabNZB')" data-prefix="Apps" data-sslname="ValidSSL" data-names='["Name","", "URL","APIKey","Interval","Timeout"]' type="button" title="Add another instance of SabNZB." class="add-new-item-button btn btn-primary"><i class="fa fa-plus"></i></button>
                                                     </div>
                                                 </td>
                                                 <td colspan="7" class="tablet-hide desktop-hide">
@@ -93,7 +93,7 @@
                                                                 {{- end}}
                                                                 <input type="text" id="Apps.SabNZB.{{$index}}.URL" onChange="showhttps($(this).val(), '#SabNZB{{$index}}SSL');" name="Apps.SabNZB.{{$index}}.URL" data-index="{{$index}}" data-app="SabNZB" class="client-parameter form-control input-sm" data-group="download" data-label="SabNZB {{instance $index}} URL" data-original="{{$app.URL}}" value="{{$app.URL}}">
                                                                 <div style="width:30px; max-width:30px;{{if not (contains $app.URL "https://")}}display:none;{{end}}" id="SabNZB{{$index}}SSL" class="input-group-addon input-sm">
-                                                                    <input type="checkbox" id="Apps.SabNZB.{{$index}}.VerifySSL" name="Apps.SabNZB.{{$index}}.VerifySSL" data-index="{{$index}}" data-app="SabNZB" class="client-parameter" data-group="download" data-label="SabNZB {{instance $index}} SSL" data-original="{{$app.VerifySSL}}" {{if $app.VerifySSL}}checked {{end}}value="true">
+                                                                    <input type="checkbox" id="Apps.SabNZB.{{$index}}.ValidSSL" name="Apps.SabNZB.{{$index}}.ValidSSL" data-index="{{$index}}" data-app="SabNZB" class="client-parameter" data-group="download" data-label="SabNZB {{instance $index}} SSL" data-original="{{$app.ValidSSL}}" {{if $app.ValidSSL}}checked {{end}}value="true">
                                                                 </div>
                                                             </div>
                                                         </div>
