+++ pkg/bindata/templates/index.html
@@ -148,12 +148,6 @@
                             <div class="navigation-item" id="template-config" style="display: none;">
 {{ template "config.html" . }}
                             </div>
-                            <div class="navigation-item" id="template-services" style="display: none;">
-{{ template "services.html" . }}
-                            </div>
-                            <div class="navigation-item" id="template-system" style="display: none;">
-{{ template "system.html" . }}
-                            </div>
                             <div class="navigation-item" id="template-starr" style="display: none;">
 {{ template "starr/index.html" . }}
                             </div>
@@ -163,14 +157,26 @@
                             <div class="navigation-item" id="template-media" style="display: none;">
 {{ template "media/index.html" . }}
                             </div>
+                            <div class="navigation-item" id="template-snapshot" style="display: none;">
+{{ template "snapshot/index.html" . }}
+                            </div>
+                            <div class="navigation-item" id="template-filewatcher" style="display: none;">
+{{ template "filewatcher.html" . }}
+                            </div>
+                            <div class="navigation-item" id="template-commands" style="display: none;">
+{{ template "commands.html" . }}
+                            </div>
+                            <div class="navigation-item" id="template-services" style="display: none;">
+{{ template "services.html" . }}
+                            </div>
                             <div class="navigation-item" id="template-triggers" style="display: none;">
 {{ template "triggers.html" . }}
                             </div>
-                            <div class="navigation-item" id="template-monitoring" style="display: none;">
-{{ template "monitoring.html" . }}
-                            </div>
                             <div class="navigation-item" id="template-integrations" style="display: none;">
 {{ template "integrations.html" . }}
+                            </div>
+                            <div class="navigation-item" id="template-monitoring" style="display: none;">
+{{ template "monitoring.html" . }}
                             </div>
                             <div class="navigation-item" id="template-metrics" style="display: none;">
 {{ template "metrics.html" . }}
@@ -181,14 +187,8 @@
                             <div class="navigation-item" id="template-configfiles" style="display: none;">
 {{ template "configfiles.html" . }}
                             </div>
-                            <div class="navigation-item" id="template-filewatcher" style="display: none;">
-{{ template "filewatcher.html" . }}
-                            </div>
-                            <div class="navigation-item" id="template-commands" style="display: none;">
-{{ template "commands.html" . }}
-                            </div>
-                            <div class="navigation-item" id="template-snapshot" style="display: none;">
-{{ template "snapshot/index.html" . }}
+                            <div class="navigation-item" id="template-system" style="display: none;">
+{{ template "system.html" . }}
                             </div>
                             <div class="navigation-item" id="template-processlist" style="display: none;">
 {{ template "processlist.html" . }}
